Moved common functionality to abstract CosmosHealthCheck class
