+++ src/Microsoft.Health.ControlPlane.CosmosDb/Health/ControlPlaneCosmosHealthCheck.cs
@@ -3,28 +3,18 @@
 // Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // -------------------------------------------------------------------------------------------------
 
-using System;
-using System.Threading;
-using System.Threading.Tasks;
-using EnsureThat;
 using Microsoft.Azure.Documents;
-using Microsoft.Extensions.Diagnostics.HealthChecks;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Microsoft.Health.CosmosDb.Configs;
+using Microsoft.Health.CosmosDb.Features.Health;
 using Microsoft.Health.CosmosDb.Features.Storage;
 using Microsoft.Health.Extensions.DependencyInjection;
 
 namespace Microsoft.Health.ControlPlane.CosmosDb.Health
 {
-    public class ControlPlaneCosmosHealthCheck : IHealthCheck
+    public class ControlPlaneCosmosHealthCheck : CosmosHealthCheck
     {
-        private readonly IScoped<IDocumentClient> _documentClient;
-        private readonly CosmosDataStoreConfiguration _configuration;
-        private readonly CosmosCollectionConfiguration _collectionConfiguration;
-        private readonly IDocumentClientTestProvider _testProvider;
-        private readonly ILogger<ControlPlaneCosmosHealthCheck> _logger;
-
         /// <summary>
         /// Initializes a new instance of the <see cref="ControlPlaneCosmosHealthCheck"/> class.
         /// </summary>
@@ -39,39 +29,16 @@ namespace Microsoft.Health.ControlPlane.CosmosDb.Health
             IOptionsSnapshot<CosmosCollectionConfiguration> namedCosmosCollectionConfigurationAccessor,
             IDocumentClientTestProvider testProvider,
             ILogger<ControlPlaneCosmosHealthCheck> logger)
+            : base(
+                    documentClient,
+                    configuration,
+                    namedCosmosCollectionConfigurationAccessor,
+                    testProvider,
+                    logger)
         {
-            EnsureArg.IsNotNull(documentClient, nameof(documentClient));
-            EnsureArg.IsNotNull(configuration, nameof(configuration));
-            EnsureArg.IsNotNull(namedCosmosCollectionConfigurationAccessor, nameof(namedCosmosCollectionConfigurationAccessor));
-            EnsureArg.IsNotNull(testProvider, nameof(testProvider));
-            EnsureArg.IsNotNull(logger, nameof(logger));
-
-            _documentClient = documentClient;
-
-            EnsureArg.IsNotNull(_documentClient, optsFn: options => options.WithMessage("Factory returned null."));
-
-            _configuration = configuration;
-            _collectionConfiguration = namedCosmosCollectionConfigurationAccessor.Get(Constants.CollectionConfigurationName);
-            _testProvider = testProvider;
-            _logger = logger;
+            CollectionConfiguration = namedCosmosCollectionConfigurationAccessor.Get(Constants.CollectionConfigurationName);
         }
 
-        public async Task<HealthCheckResult> CheckHealthAsync(HealthCheckContext context, CancellationToken cancellationToken = default)
-        {
-            try
-            {
-                // Make a non-invasive query to make sure we can reach the data store.
-
-                await _testProvider.PerformTest(_documentClient.Value, _configuration, _collectionConfiguration);
-
-                return HealthCheckResult.Healthy("Successfully connected to the data store.");
-            }
-            catch (Exception ex)
-            {
-                _logger.LogWarning(ex, "Failed to connect to the data store.");
-
-                return HealthCheckResult.Unhealthy("Failed to connect to the data store.");
-            }
-        }
+        public override CosmosCollectionConfiguration CollectionConfiguration { get; }
     }
 }
